chore: renames IngressGatewaySpec type to GatewaySpec (#1144)
Original name is confusing as it does not actually define anything
related to the ingress. It is configuring Istio Gateway resource which
is, in case of Model Serving, attached to ingress service.

This change makes the change of the golang type, but not publicly
exposed `ingressGateway` config defined for KServe component.

(cherry picked from commit e80dcbe8b28d5dcc77284498466055532a20b80c)