fix(feature): ensures all new resources are created in one reconcile (#1159)
Manifests used by Feature to create cluster resources can be either
defined as a single resource per file or grouped together and separated
using the `---` delimiter. Even though the former approach is currently
preferred across the codebase, the latter is used to create Envoy filters
for the KServe setup.

When applying resources not yet present in the cluster, the original
implementation returns after creating the first resource, leaving the rest
unprocessed. They are applied in the next reconcile loop, though again
on a one-per-reconcile event basis. This is not a bug per se, as it will
eventually result in creating all resources; however, it is suboptimal.

This change ensures all resources defined in a single YAML manifest,
separated using `---`, are processed in a single go.

(cherry picked from commit 959643d01a424d87d36e272e88edd614a589400e)