Update golang.org/x/net to v0.33.0
JIRA: https://issues.redhat.com/browse/RHOAIENG-17203

This change is to update to a version that isn't flagged as affected
by CVE-2024-45338.

While we know that this operator doesn't call the vulnerable code
paths (hooray for `govulncheck`), some scanning tools are not yet
smart enough to be able to discern that. So it's preferable just to
update.

Go advisory: https://pkg.go.dev/vuln/GO-2024-3333

This is the 2.16 backport of commit 855a8b4.